Create separate file for model option
